new api and fix for the tournament kill leader list

+++ Fanview.API/Model/ViewModels/Kills.cs
@@ -8,7 +8,7 @@ namespace Fanview.API.Model.ViewModels
     public class Kills
     {
         public string playerName { get; set; }
-        public float playerId { get; set; }
+        public int playerId { get; set; }
         public int kills { get; set; }
         public int timeSurvived { get; set; }
         public int teamId { get; set; }
